Moving tests back into position.
